+++ src/api/BdmsContextExtensions.cs
@@ -202,6 +202,9 @@ public static class BdmsContextExtensions
            .RuleFor(o => o.PrecisionLocationY, f => f.PickRandom(Enumerable.Range(0, 10)))
            .RuleFor(o => o.PrecisionLocationXLV03, f => f.PickRandom(Enumerable.Range(0, 10)))
            .RuleFor(o => o.PrecisionLocationYLV03, f => f.PickRandom(Enumerable.Range(0, 10)))
+           .RuleFor(o => o.TotalDepthTvd, _ => null)
+           .RuleFor(o => o.TopBedrockFreshTvd, _ => null)
+           .RuleFor(o => o.TopBedrockWeatheredTvd, _ => null)
            .RuleFor(o => o.Observations, _ => new Collection<Observation>())
            .FinishWith((f, o) => { o.Name = o.OriginalName; });