+++ src/api/BoreholeGeometry/XYZFormat.cs
@@ -20,7 +20,7 @@ internal sealed class XYZFormat : IBoreholeGeometryFormat
     public IList<BoreholeGeometryElement> ReadCsv(IFormFile file, int boreholeId)
     {
         using var reader = new StreamReader(file.OpenReadStream());
-        using var csv = new CsvReader(reader, CsvConfigHelper.CsvConfig);
+        using var csv = new CsvReader(reader, CsvConfigHelper.CsvReadConfig);
 
         var data = csv.GetRecords<Geometry>();
         return ToBoreholeGeometry(data, boreholeId);